Extract styles into separate build step
