+++ proxy.Dockerfile
@@ -11,6 +11,21 @@ RUN --mount=type=bind,source=proxy/js/ui.js,target=ui.js.tmpl \
     | template --configuration - --format yaml --template ui.js.tmpl \
     > /build/ui.js
 
+FROM node:22-alpine as build-styles
+
+WORKDIR /build
+
+RUN npm install yaml
+
+RUN --mount=type=bind,source=proxy/js/styles.mjs,target=styles.mjs \
+  --mount=type=bind,source=features/train_protection.yaml,target=train_protection.yaml \
+  --mount=type=bind,source=features/speed_railway_signals.yaml,target=speed_railway_signals.yaml \
+  --mount=type=bind,source=features/electrification_signals.yaml,target=electrification_signals.yaml \
+  --mount=type=bind,source=features/signals_railway_signals.yaml,target=signals_railway_signals.yaml \
+  node /build/styles.mjs \
+
+RUN node styles.mjs
+
 FROM nginx:1-alpine
 
 COPY proxy/proxy.conf.template /etc/nginx/templates/proxy.conf.template
@@ -21,3 +36,6 @@ COPY proxy/image /etc/nginx/public/image
 
 COPY --from=build-ui \
   /build/ui.js /etc/nginx/public/js/ui.js
+
+COPY --from=build-styles \
+  /build /etc/nginx/public/style
